# postman/specs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SEAPEDIA API",
        "description": "Campus marketplace API — multi-role (buyer/seller/driver) with Sanctum Bearer auth.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local dev (Sail)"
        }
    ],
    "paths": {
...
        "/api/v1/admin/promos": {
            "get": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "List all promos",
                "operationId": "9736bca2a907bf55b94382f11d5b4833",
                "responses": {
                    "200": {
                        "description": "List of promos"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "Generate a new promo code",
                "operationId": "76c447179c40fcb594320aaa72d12853",
                "responses": {
                    "201": {
                        "description": "Promo created"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/promos/{promo}": {
            "get": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "View a single promo",
                "operationId": "e91b988dc87bb695a4d5c50418766dc2",
                "responses": {
                    "200": {
                        "description": "Promo detail"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/vouchers": {
            "get": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "List all vouchers",
                "operationId": "04deb7ad743db63cfdcdf06faf534be2",
                "responses": {
                    "200": {
                        "description": "List of vouchers"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "Generate a new voucher code",
                "operationId": "22e02aad204fd13605832fc36bbfd71d",
                "responses": {
                    "201": {
                        "description": "Voucher created"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/vouchers/{voucher}": {
            "get": {
                "tags": [
                    "Admin Discounts"
                ],
                "summary": "View a single voucher",
                "operationId": "3b8dcb5ce2d9f021cd2fbf599216cfa9",
                "responses": {
                    "200": {
                        "description": "Voucher detail"
                    },
                    "403": {
                        "description": "Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate and mint a Sanctum token",
                "description": "Mint a Sanctum token for non-SPA `/api/v1` consumers (§5.0). A\nsingle-role user (or admin) gets a role-scoped token immediately; a\nmulti-role user gets an unscoped token and must call `/role/select`.",
                "operationId": "a30aac1b7501f4f24e85306d1c4cec2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token issued (role-scoped or pending role selection)"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    },
                    "429": {
                        "description": "Too many login attempts — throttled"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current Sanctum token",
                "description": "Revoke the token used for this request (§5.0 \"logout invalidates\n     * token\" — proven by the same token being rejected on the next call).",
                "operationId": "b76198bd1b4798bab3a0b97a4f63b6a0",
                "responses": {
                    "200": {
                        "description": "Token revoked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/buyer/cart": {
            "get": {
                "tags": [
                    "Buyer Cart"
                ],
                "summary": "Get the buyer's cart summary",
                "operationId": "e962abb74af0a54cdeb560d2427b5393",
                "responses": {
                    "200": {
                        "description": "Cart with its store and items"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/cart/items": {
            "post": {
                "tags": [
                    "Buyer Cart"
                ],
                "summary": "Add a product to the cart (single-store guard, §5.8)",
                "operationId": "8bec063fa9c4b1a2d6bacce4db95df3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "replace": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added"
                    },
                    "422": {
                        "description": "Cart holds items from a different store"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/cart/items/{item}": {
            "put": {
                "tags": [
                    "Buyer Cart"
                ],
                "summary": "Update a cart item quantity",
                "operationId": "576a466f2486c51190a6c2975c4495eb",
                "responses": {
                    "200": {
                        "description": "Item updated"
                    },
                    "403": {
                        "description": "Not the owner"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Buyer Cart"
                ],
                "summary": "Remove a cart item",
                "operationId": "9ca7b885f6661605798aa217015f4a39",
                "responses": {
                    "204": {
                        "description": "Item removed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/buyer/cart/clear": {
            "post": {
                "tags": [
                    "Buyer Cart"
                ],
                "summary": "Clear the cart",
                "operationId": "f7741bcd6a66e5bcb8d1b8601a3a1eb9",
                "responses": {
                    "204": {
                        "description": "Cart cleared"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/role/select": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Pick the active role and mint a role-scoped token",
                "description": "Mint a role-scoped token and revoke the pending one used to call this\nendpoint (§5.0) — only one live token per active-role session.",
                "operationId": "20761e7f00c729252b42c9e9acd0fe76",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "buyer",
                                            "seller",
                                            "driver"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role-scoped token issued"
                    },
                    "422": {
                        "description": "User does not own this role"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum token — obtain from POST /api/v1/login",
                "scheme": "bearer"
            }
        }
    },
...
}